+++ com.unity.shadergraph/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [6.8.0] - 2019-XX-XX
+### Added
+- Color Mode improves node visibility by coloring the title bar by Category, Precision, or custom colors
 
 ## [6.7.0] - 2019-05-21
 ### Added